header parameter

+++ app.py
@@ -1,6 +1,9 @@
 from flask import Flask, request
 from flask_restful import Resource, reqparse, Api
-from flask_httpauth import HTTPBasicAuth
+# from flask_httpauth import HTTPBasicAuth
+import sys
+import logging
+
 
 app = Flask(__name__)
 api = Api(app)
@@ -8,7 +11,11 @@ app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///base.db'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 app.config['PROPAGATE_EXCEPTIONS'] = True
 
-auth = HTTPBasicAuth()
+# auth = HTTPBasicAuth()
+
+
+app.logger.addHandler(logging.StreamHandler(sys.stdout))
+app.logger.setLevel(logging.ERROR)
 
 
 from base import data, db
@@ -17,15 +24,15 @@ app.app_context().push()
 db.create_all()
 
 # Auth_Data
-USER_DATA = {
-    "username":"password"
-}
+# USER_DATA = {
+#     "username":"password"
+# }
 
-@auth.verify_password
-def verify(username, password):
-    if not (username and password):
-        return False
-    return USER_DATA.get(username) == password
+# @auth.verify_password
+# def verify(username, password):
+#     if not (username and password):
+#         return False
+#     return USER_DATA.get(username) == password
 
 
 class data_List(Resource):
@@ -35,15 +42,23 @@ class data_List(Resource):
     parser.add_argument('last_seen', type=str, required=True, help='last_seen of the data_Item')
     parser.add_argument('timestamp', action='append', required=True, help='timestamp list of the data_Item')
     
-    @auth.login_required
+    # @auth.login_required
     def get(self, data_Item_Id):
+        ck = self.check()
+        if ck == "private":
+            return {'message': 'data invalid!'}
+        
         item = data.find_by_id(data_Item_Id)
         if item:
             return item.json()
         return {'Message': 'ID = {} is not found'.format(data_Item_Id)}
     
-    @auth.login_required
+    # @auth.login_required
     def post(self, data_Item_Id):
+        ck = self.check()
+        if ck == "private":
+            return {'message': 'data invalid!'}
+        
         if data.find_by_id(data_Item_Id):
             return {' Message': 'Data Item with ID = {} already exists'.format(data_Item_Id)}
         args = data_List.parser.parse_args()
@@ -51,8 +66,12 @@ class data_List(Resource):
         item.save_to()
         return item.json()
     
-    @auth.login_required
+    # @auth.login_required
     def put(self, data_Item_Id):
+        ck = self.check()
+        if ck == "private":
+            return {'message': 'data invalid!'}
+        
         args = data_List.parser.parse_args()
         print(args.timestamp)
         item = data.find_by_id(data_Item_Id)
@@ -67,21 +86,52 @@ class data_List(Resource):
         item.save_to()
         return item.json()
     
-    @auth.login_required    
+    # @auth.login_required    
     def delete(self, data_Item_Id):
+        ck = self.check()
+        if ck == "private":
+            return {'message': 'data invalid!'}
+        
         item  = data.find_by_id(data_Item_Id)
         if item:
             item.delete()
             return {'Message': 'Data Item with ID = {} has been deleted from records'.format(data_Item_Id)}
         return {'Message': 'ID = {} is already not on the list'.format(data_Item_Id)}
-
-class All_data(Resource):
-    @auth.login_required
-    def get(self):
-        return {'Data': list(map(lambda x: x.json(), data.query.all()))}
     
+    def check(self):
+        if "key" in request.headers:
+            if request.headers['key'] != "secret":
+                return 'private'
+        else:
+            return 'private'
+
+        
+        
+class All_data(Resource):
+    def check(self):
+        if "key" in request.headers:
+            if request.headers['key'] != "secret":
+                return 'private'
+        else:
+            return 'private'
+    # @auth.login_required
+    def get(self):
+        
+        ck = self.check()
+        if ck == "private":
+            return {'message': 'data invalid!'}
+
+        return {'Data': list(map(lambda x: x.json(), data.query.all()))}
+
+
 api.add_resource(All_data, '/')
 api.add_resource(data_List, '/<data_Item_Id>')
+
+# @app.route('/')
+# @auth.login_required
+# def get():
+#     return {'Data': list(map(lambda x: x.json(), data.query.all()))} 
+
 
 if __name__=='__main__':
     app.run(debug=True)
